Resource based authorization (create custom service to authorize user)

+++ Restaurants.Application/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantCommandHandler.cs
@@ -2,8 +2,10 @@
 using MediatR;
 using Microsoft.Extensions.Logging;
 using Restaurants.Application.Restaurants.Command.CreateRestaurant;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 using System;
 using System.Collections.Generic;
@@ -14,7 +16,8 @@ using System.Threading.Tasks;
 namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant
 {
     public class DeleteRestaurantCommandHandler(ILogger<CreateRestaurantCommandHandler> logger,
-        IMapper mapper, IRestaurantsRepository restaurantsRepository) : IRequestHandler<DeleteRestaurantCommand>
+        IMapper mapper, IRestaurantsRepository restaurantsRepository,
+        IRestaurantAuthorizationService restaurantAuthorizationService) : IRequestHandler<DeleteRestaurantCommand>
     {
         public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
         {
@@ -22,6 +25,9 @@ namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant
             var restaurant = await restaurantsRepository.GetById(request.Id);
             if (restaurant == null)
                 throw new NotFoundException(nameof(Restaurant),request.Id.ToString());
+
+            if (!restaurantAuthorizationService.Authorize(restaurant, ResourceOperations.Delete))
+                throw new ForbidException(); 
 
             await restaurantsRepository.Delete(restaurant);
